Move add-token-counts to llms directory, adjust getPageTitle, add comments
